# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server for Article API",
    "version": "1.0.0",
    "title": "Node-RESTful-Swagger",
    "termsOfService": "",
    "contact": {
      "email": "vkurniawan@binar.co.id"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Article",
      "description": "Articles Resources"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/articles": {
      "get": {
        "tags": [
          "Article"
        ],
        "summary": "List all available articles",
        "description": "described hit",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Article"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Article"
        ],
        "summary": "Add new article",
        "description": "Create new article",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Article object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Article"
            }
          }
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully create new article",
            "schema": {
              "$ref": "#/definitions/Article"
            }
          },
          "400": {
            "description": "Failed to create new article",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to create new article"
                }
              }
            }
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "tags": [
          "Article"
        ],
        "summary": "List an article",
        "description": "described hit",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of article to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Article"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Article": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "title": {
          "type": "string",
          "example": "Lorem Ipsum"
        },
        "body": {
          "type": "string",
          "example": "Lorem Ipsum Dolor Et Amet"
        },
        "approved": {
          "type": "boolean",
          "default": false
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Article API",
    "url": "http://article.io"
  }
}